[FEATURE]#246 : MessageRoomId Long -> Int 변경

+++ core/model/src/main/kotlin/com/bff/wespot/model/message/response/MessageRoom.kt
@@ -3,14 +3,14 @@ package com.bff.wespot.model.message.response
 import java.time.LocalDateTime
 
 data class MessageRoom(
-    val messageRoomId: Long = -1,
+    val messageRoomId: Int = -1,
     val name: String = "",
     val thumbnail: String = "",
     val isBookmarked: Boolean = false,
     val messageDetails: List<MessageDetail> = listOf(),
 ) {
     data class MessageDetail(
-        val id: Long = -1,
+        val id: Int = -1,
         val createdAt: LocalDateTime? = LocalDateTime.MIN,
         val content: String = "",
         val isReceived: Boolean = true,
